Conflicts:
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.exe
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.ilk
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.pdb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp.sdf
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp.v11.suo
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/CL.read.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/CL.write.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Config.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Customize.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/EnemyObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/EnemyProjectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Engine.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Game.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/GameObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/GameObjectManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/InputManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/MainMenu.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/MovingGameObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Options.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/PlayerObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/PlayerProjectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Projectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Spawner.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Sprite.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SpriteManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/StateManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SuitEmUp.lastbuildstate
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SuitEmUp.log
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/TitleScreen.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Truck.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Utils.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/cl.command.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.command.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.read.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.write.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/main.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/vc110.idb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/vc110.pdb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Score.cpp
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/SuitEmUp.vcxproj.filters

+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/EyeCandy.h
@@ -21,7 +21,7 @@ public:
 	void BloodCreator(char* p_type, sf::Vector2f p_position, sf::Vector2f p_direction);
 	void PictureCreator(sf::Sprite* p_sprite, sf::Vector2f p_position, float p_rotation);
 
-	void BoomWoshCreator(sf::Vector2f p_position, sf::Vector2i p_destination);
+	void BoomWoshCreator(sf::Vector2f p_position, sf::Vector2f p_destination);
 	void TextCreator(sf::Text p_text, sf::Vector2f p_position);
 
 	void Update(float deltatime);
@@ -36,7 +36,7 @@ private:
 		float m_duration;
 		sf::Vector2f m_direction;
 		sf::Vector2f m_position;
-		sf::Vector2i m_destination;
+		sf::Vector2f m_destination;
 		sf::Vector2f m_velocity;
 		sf::Vector2f m_acceleration;
 		sf::Vector2f m_aacceleration;
